feat: New generated content (#424)
Update neutron schema resolution

Neutron treats subnetpool.XXX_prefixlen everywhere as integers including
the database (it is even described like that in neutron api-ref itself).
For some unknown reason it returns it as a string though. Maybe because
in the apidef it is defined like `validate: "type:non_negative",
convert_to: convert_to_int`. Modify api_def processing not to look at
validate OR convert_to, but use convert_to results in update of the
schema.


Changes are triggered by https://review.opendev.org/925868

+++ openstack_sdk/src/api/network/v2/policy/packet_rate_limit_rule/list.rs
@@ -41,13 +41,13 @@ pub struct Request<'a> {
     /// /v2.0/policies/{policy_id}/packet_rate_limit_rules API
     ///
     #[builder(default)]
-    max_burst_kpps: Option<f32>,
+    max_burst_kpps: Option<i32>,
 
     /// max_kpps query parameter for
     /// /v2.0/policies/{policy_id}/packet_rate_limit_rules API
     ///
     #[builder(default)]
-    max_kpps: Option<f32>,
+    max_kpps: Option<i32>,
 
     /// policy_id parameter for
     /// /v2.0/policies/{policy_id}/packet_rate_limit_rules/{id} API
